requirements.txt
streamlit
pandas
numpy
scikit-learn
joblib
paho-mqtt